% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighestMedianRules.R
\docType{data}
\name{elec2012}
\alias{elec2012}
\title{Example voting profile}
\format{An object of class \code{matrix} with 10 rows and 7 columns.}
\usage{
elec2012
}
\description{
Matrix containing the shares of grades (in -2:4) of candidates on each row. Taken from Balinski & Laraki (2016), "Majority Judgment vs Majority Rule" p. 14. Grades given by a non-representative sample of 737 voters for the 2012 French presidential election.
}
\keyword{datasets}
